refactor and split out flatgrid data to hooks
